// src/modules/product/enum/ProductInsertTestId.enum.ts
export enum ProductInsertTestIdEnum {
  PRODUCT_INSERT_CONTAINER = 'PRODUCT_INSERT_CONTAINER',
  PRODUCT_INPUT_NAME = 'PRODUCT_INPUT_NAME',
  PRODUCT_INPUT_IMAGE = 'PRODUCT_INPUT_IMAGE',
  PRODUCT_INPUT_PRICE = 'PRODUCT_INPUT_PRICE',
  PRODUCT_INPUT_SELECT = 'PRODUCT_INPUT_SELECT',
  PRODUCT_INPUT_WEIGHT = 'PRODUCT_INPUT_WEIGHT',
  PRODUCT_INPUT_LENGTH = 'PRODUCT_INPUT_LENGTH',
  PRODUCT_INPUT_WIDTH = 'PRODUCT_INPUT_WIDTH',
  PRODUCT_INPUT_HEIGHT = 'PRODUCT_INPUT_HEIGHT',
  PRODUCT_INPUT_DIAMETER = 'PRODUCT_INPUT_DIAMETER',
  PRODUCT_BUTTON_CANCEL = 'PRODUCT_BUTTON_CANCEL',
  PRODUCT_BUTTON_INSERT = 'PRODUCT_BUTTON_INSERT',
}